servicio-fe/requirements.txt
Flask==3.1.2
google-cloud-firestore==2.21.0
gunicorn==23.0.0
